Split out `ShearedButton` from `ShearedToggleButton`
